Change default error messages for validations
The playtime section of Iteration B2 Unit Testing Models includes
a task to change an error message associated with one validation on the
Product model.

I have changed the error messages for:
- The presence helper on the description attribute
- The uniqueness helper on the title attribute

There is a simpler way to do it however some links suggest doing it this
way using locales is the accepted rails way [1], and we use this
approach in af2.

Trello: https://trello.com/c/Px7y2B7D