Add button 'Add to Cart'
In order to add items to the cart a button_to has been added to the
homepage, and styled.

Following rails convention this goes to #create, in the
LineItemsController where a new LineItem is created, then built with the
product id, which is taken from params. The build method is actually an
alias for new, which works too [1]
https://api.rubyonrails.org/v7.0.3/classes/ActiveRecord/Associations/
CollectionProxy.html#method-i-build

Flash notices inform the user whether the LineItem has been successfully
save, they are then redirected to the cart page.

One test has been added, the flash messages could also be tested.
The set_cart method from the CurrentCart concern could be renamed as
there is another set_cart method in the CartsController.