Handle multiple quantities of a product
This commit introduces a method in the Cart model which allows us to
increment the quantity attribute of LineItem, so that when multiples of
the same product are added to a Cart, we show that rather than one
LineItem for each product added.

A migration has cleaned up the database removing the individual entries
and summing them before adding back an entry with a correct quantity.

The view has also been updated.