Handle potential exception for invalid cart id
Iteration E2
This commit handles the ActiveRecord::RecordNotFound exception
which would be raised if a user entered a cart id which does not exist.

A private method has been added in the CartsController to redirect the
user to the homepage and display a flash message.

The cart params are not longer permitted in the LineItemsController,
and the LineItemsControllerTest has also been updated