Add a further test for Iteration E2 Error handling
Iteration E2 adds a handler method to rescue from an
ActiveRecord::RecordNotFound error which could be caused by a user
entering an invalid cart id.

I have added a further test for the workflow where an invalid cart
id is entered, the exception is raised and handled by redirection to
the homepage and a flash message.

The test tests all three things, it felt right to test the redirection
and the flash message in the same test as this is similar to the example
in the rails guide [1] however I am less sure that asserting
the exception is raised should be in the same test.

Also the fixture is created in the test, I wanted to create an
invalid_cart fixture but I couldn't get the test to find the required
id.

[1] https://guides.rubyonrails.org/testing.html#testing-flash-notices