Add ability to empty the cart
Iteration E3 adds a delete button on the homepage and functionality to
delete a cart.

The display of the cart page is updated to show the line_items in a
table, the total price of each line item, and of the whole cart is
calculated using two methods named total_price, one in the Carts Model
and one in LineItem.

Styling is updated.

Some testing is done in the book, the delete action in the CartsController
is tested, however the model methods are not tested, unit tests will be
added later.

An extra system test is broken by the change in text. System testing is
added later in the book, however I might fix some of the system tests
ahead of that.