Add highlighting to improve UX when Adding to cart
When the user clicks the Add to cart button they now remain on the
same page, without a refresh due to a previous commit adding AJAX.

To make it clearer to the user that an item has been added to the cart
css animation has been added to highlight items for one second after
they have been added to the cart.

The associated test has been updated.