Add ActionCable to update product price
ActionCable allows realtime updates without a page refresh
or user interaction. Here it is set up only for use with Products.

The test file has been added, a card has been added to the low
priority lane to add testing. As we do not use ActionCable in our
applications I am leaving the testing for now.