Add a helper test for Iteration F4
This adds a test for the render_if helper which conditionally
displays the cart.

I have used assert_nil to assert that the partial is not displayed
when the cart is present but has no line_items. This seemed like a
good option as nothing else is rendered when the condition in the
method returns false. There is a similar test in af2 for the loop helper.