Add unit tests for add_product method in Cart model
This is a refactor of earlier work adding Unit tests for the
add_product method in Cart model.
Some tests have been removed as they were just understanding how fixtures
work, in the near future FactoryBot will be added and will replace
fixtures.

A change to one of the controller tests will also be made as the fixures
are all loaded for each test any changes mean updates to unrelated tests
are sometimes required.