Remove the counter for store visits
This was added as a playtime activity in Task D. The addition of
ApplicationCable has broken this functionality, which does not seem
to be something that is carried on outside of that extension activity.

A it does not appear to add much value it is better to remove it than
spend time fixing it.