Add scaffolding for Orders
This commit adds all the scaffolding files for Orders, as well as the
two migrations used to create the orders table and add a reference
to it in the line items table.

An enum has been added to the Order model so that we can handle different
types of payment as integers rather than put strings in the database.