This commit adds functionality to create an Order
The code from Iteration G1 is added which includes:
- Functionality to add an Order using a form
- Updates to the Orders#create action
- Validation on the Order model
- Rendering of the view corresponding to Orders#new
- Styling of the order form
- Updates to a controller test

I have added additional tests to the Orders model, it still
requires testing for the ensure_cart_isnt_empty method, however I
need to resolve a DuplicateDefinition error with the Orders factory.
The orders fixtures have been left in for now but will be removed when
all the tests using them have been updated.

System tests and further controller tests are going to be added in a
later commit.