Add additional jsx files for pay type options
This Commit adds additional components for each of the different pay
types. When an option is selected the relevant component is rendered,
providing a dynamic display of the required form fields.

A controller method is added to handle params, though not fully
integrated until iteration I.

The code does lead to a warning about the Reach render() method being
deprecated as well as guidance about using autocomplete attributes for
form fields. Trello cards have been spun out for both.