Update Orders system test to use FactoryBot
The Test in the book for the select does not work, it should be able
to work with the Javascript using the selenium driver, locally I have
added some Javascript to check this. The output of the Javascript is not
being found by the test, though you can see the output on the test
screenshot.
The html from the React component does not output in the html using
puts page.html or appear in the screenshot.