Add functionality to send emails
Iteration I1 uses ActionMailer to send emails to customers when they
place an order.
The application has been configured the same configuration for all
environments, this might be changed later.

The sensitive information has been put in a .env file using the
dotenv-rails gem.

The text templates have been removed so that we just use the html
templates.