Add BCrypt for User authentication
The BCrypt gem has been added and used in the conventional way by adding
has_secure_password to the User model. Validations have been added to
the User model and tested.

A factory for Users uses BCrypt to generate passwords. The default cost
has been reduced from the default of 10 to 4, which makes the tests run
a slightly faster.