This covers Iteration J2 Autenticating Users
The management of Users is split between a SessionsController and
an AdminController. The former manages the users log in and out, while
the AdminController for now just displays information.

The SessionsController allows a user to log in with a secure password
and sets their user id in the session as :user_id. They are redirected
to the Admin page which for now just shows them the number of orders.

The SessionsController also manages logging out.

Routes have been adapted to be a bit more user friendly.
I have added more to the SessionsControllerTest than is in the book,
the test for logging in and out is within one test as it is the same
data flow, with FactoryBot I would have to duplicate the set up of a
logged in user, I did try this in a private method, though at present it
is not needed in any other tests so seemed like duplication in another
place.

System tests are needed though I will add at the end of the Chapter.