Restrict access for users who are not logged in
This commit adds a before_action in ApplicationController to filter out
users who are not logged in. The before_action is skipped for
controller actions which need to be accessible to non logged in users.

An AuthenticationHelpers module has been added to the test_helper and
included in both ActionDispatch::IntegrationTest and
ActionDispatch::SystemTestCase so that users are logged in as
necessary for the tests without having to setup to log the user in in
each test requiring it.