Refactor SessionsControllerTest
Now that we have AuthenticationHelpers for logging the user in and out
through out the test suite where needed, the tests for the login and
log out functionality are not needed as they were.

They have been updated and re named to focus on the redirection and
flash notice / alert the User sees when they log in and log out.