Raise an exception when the last user is deleted
This change makes sure that the last User cannot be deleted.
It uses a method in the User model, which is rescued in the
UsersController

I would like to test this, though at present have not found a way of
doing this, or understanding exactly why the test to destroy a User
in the UsersControllerTest still passes.