Add locales to the application
This iteration adds locales for :en and :es, I already had set en-GB as
the default locale to ensure that currency was displayed as pounds
sterling not US dollars, so I have kept this as the default, though
for the sake of keeping in line with the book I have made :es and :en
available_locales.

A scope has been added to the routes so that only the store area is
internationalized.