Add translations for the Store
Translations have been added to en-GB and es for the home page, some
tests will need updating.
It does not seem necessary to add a locale file for :en as this is just
used to keep in line with the book, so a fallback will be configured in
a subsequent commit.