Changes following PR review
The wrong Pluralize method was used, this has been amended, we are not
adding Internationalization to the Admin section though it might be
nicer to add a translation key instead of having text in the view,
incase this method gets used anywhere else.

A System test has been added for the Admin page, we should also be
testing the links including the logout. These will be added later on.

The loading of fixtures has been commented out in the test_helper as
otherwise it confuses the test, if fixtures are loaded we get those as
well as any FactoryBot instances. All the tests need fixtures removing.