Upgrade Ruby to 2.7.0 to install i18n-js
Following the instructions in the Book for Iteration K, install of
i18n-js fails, [1].

Following discussion with Clare Reid we decided it was worth upgrading
to use Ruby 2.7.0. This commit does that and the additional required
changes to the Gemfile from the install of i18n-js 4.0.1.

The github repo for i18n-js have been followed rather than those in
the book as config.I18n::JS::Middleware is still giving an Uninitialized
Constant error, however the instructions on the repo [2] appear to work,
and get the expected output from the i18n CLI, for example `i18n check`
lists missing and extraneous translations.
The frontend folder is auto generated

[1]ERROR:  Error installing i18n-js:
	The last version of glob (>= 0) to support your Ruby & RubyGems was 0.2.0.
  Try installing it with `gem install glob -v 0.2.0` and then running
  the current command again
	glob requires Ruby version >= 2.7.0. The current ruby version is 2.6.5.114.

[2] https://github.com/fnando/i18n-js