This commit install the JavaScript library i18n-js
This library is referenced in the repo for the gem i18n-js. [1]
under 'Integrating with your front end'

The library has been installed and added to aplication.js, not in the
same way as the book, which says to use `require`.

The dynamic script tags in application.html.erb are the same as in the
book. The initialization of the translations follows the i18n-js JavaScript
library [2].

I have changed the yml config output file path for the gem, as it wasn't
working, but now does enable access to translations in the browser.

The next step is to work out how to call the translations from within
the React files.
[1] https://github.com/fnando/i18n-js
[2] https://github.com/fnando/i18n