Update the React translations
This adds some translations for the dynamic payment type selector,
which is in the React code.

The translations come through, however it is necessary to stay in the
same locale through one cycle of adding products to cart, creating an
order and checking out. If any changes are made to the locale through
the URL during that cycle the locale changes at a further point.

We could probably generate less of the translations files, we are currently
generating all the activerecord errors, number translations and
error translations. The translations seem to
use the frontend folder through the libraries own internal config,
those files have not all been included, though for clarity I might commit
them later.