Display the notice from the UsersController conditionally
At the moment this notice displays on the index action where it does
not contain any text, see Trello card.

This change only displays the notice if it contains text.
It doesn't feel like the most elegant solution.

I would like to replace the respond_to block in the UsersController with
a flash as respond_to has been extracted to a gem in Rails 6.