Conditionally display notices in Products show view
This makes sure that the user only sees a notice when there is information
for them to see.
Previously the notice was rendered whether or not it contained text.

The ProductsControllerTest has been updated so that it asserts against
the flash notice every time it is displayed. There could be a better
way to handle this, but fixing the display first seems sensible.

No system tests have been added as there are already system tests for
the user interactions involving those notices.