Remove display of notice in Users Show view
This removes the display of the notice in the show view for Users, as
we do not currently display any notices in this view, or redirect to it
from other actions.

The controller and system tests have been updated so that we can make
sure we are displaying notices in the correct place. This commit could
be squashed into the commit were notices are conditionally displayed in
app/views/users/index.html.erb

There should be System tests for when the user triggers validation
errors, another card has been spun out for that.