Conditional Rendering of notice for LineItems Show
This updates the display of the show view for line items, and fixes
an error in the LineItemsController where it was not finding
able to create a new LineItem by filling in the new form and posting
to LineItemsController #Create, because it could not fine the product_id
in the params.

This wasn't caught before as I didn't have time to test everything I
wanted to test.