Fix the Create action of LineItemsController
I had changed the params, as I was looking at creating a new line item
using the form, when logged in as an Admin user, and not checked the
Add to cart functionality.

This makes the Add to cart functionality work, but does break the
creation of line items by an admin.